requirements.txt
pandas
scikit-learn
numpy
seaborn 
matplotlib
flask
ipykernel
xgboost
python-box
pyYAML
ensure
joblib
types-PyYAML
Flask-Cors
-e .